% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdata.R
\name{fitRegressionModels}
\alias{fitRegressionModels}
\title{Fitting regression models for each operating system from osusers dataset}
\usage{
fitRegressionModels()
}
\value{
list of the regression models of regression models
}
\description{
Fitting regression models for each operating system from osusers dataset
}
\details{
For each variable of the OS popularity is built a regression model using \code{lm}
function.
}
\seealso{
\code{lm} \code {apply}
}

