% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdata.R
\name{getForecastingInterval}
\alias{getForecastingInterval}
\title{Building dataset of monthes for forecasting}
\usage{
getForecastingInterval(a_Year, a_Month)
}
\arguments{
\item{a_Year}{a character value of the last year in the date list for forecasting in form "YYYY" (for example, "2017", "2018" etc.)}

\item{a_Month}{a character value of the last month in the date list for forecasting in form "MM" (for example, "01", "07", "10" etc.)}
}
\value{
a data frame containg the dates for futher forecasting in format "YYYY-MM-01"
}
\description{
Building dataset of monthes for forecasting
}
\details{
Building a dataset containing monthes later the greatest date from the osusers dataset
to provide a forecasting
}

