% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdata.R
\name{getPredictions}
\alias{getPredictions}
\title{Building predictions for the number of monthes}
\usage{
getPredictions(a_Year, a_Month)
}
\arguments{
\item{a_Year}{a character value of the last year in the date list for forecasting in form "YYYY" (for example, "2017", "2018" etc.)}

\item{a_Month}{a character value of the last month in the date list for forecasting in form "MM" (for example, "01", "07", "10" etc.)}
}
\value{
a list of the predictions for given monthes
}
\description{
Building predictions for the number of monthes
}
\details{
This function uses possibilities of some other function to give predictions of operating systems popularity indexis
for the month after the latest date from the osusers dataset to the month, which is given within the parameters of the
function. The function builds a number iof the regression models, each for the given within the dataset operating
system. Then it defined a list of dates for the forecasting and makes predictions.
Each prediction in the list contains the predicted value and the upper and lower values of the 95%-confidence interval
}
\seealso{
\code{fitRegressionModels} \code{getForecastingInterval} \code{lapply}
}

