% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osdata.R
\name{getPredValues}
\alias{getPredValues}
\title{Extract the predicted values}
\usage{
getPredValues(a_Year, a_Month)
}
\arguments{
\item{a_Year}{a character value of the last year in the date list for forecasting in form "YYYY" (for example, "2017", "2018" etc.)}

\item{a_Month}{a character value of the last month in the date list for forecasting in form "MM" (for example, "01", "07", "10" etc.)}
}
\value{
a dataset of the predicted values
}
\description{
Extract the predicted values
}
\details{
This function uses predicted value, taken from the \code{getPredictions} function for given time period ended with a_Year
and a_Month. Then extract it the predicted values from the predictions.
}
\seealso{
\code{fitRegressionModels} \code{getForecastingInterval} \code{lapply}
}

